+++ zis-math/src/main/java/module-info.java
@@ -1,3 +1,4 @@
 module zis.math {
-    exports org.schlunzis.zis.math;
+    exports org.schlunzis.zis.math.linear;
+    exports org.schlunzis.zis.math.analysis;
 }
